requirements.txt
fastapi
uvicorn
httpx
yfinance
pandas
pydantic
pydantic-settings
python-dotenv
fastapi-cache2[redis]
structlog
tenacity
redis